fixing some warnings
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@775 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ src/test/test_num_spline.inl
@@ -62,7 +62,7 @@ namespace test
 namespace num_spline
 {
 
-std::string doubleToString(double iX, size_t iDigitsMantissa, size_t iDigitsExponent)
+std::string doubleToString(double iX, int iDigitsMantissa, int iDigitsExponent)
 {
 	std::ostringstream buffer;
 	buffer << std::scientific << std::showpos << std::setprecision(iDigitsMantissa - 1) << iX;
